Basic limit validation working for benchtest params

+++ src/plugins/freeems/freeems_gui_handlers.c
@@ -246,7 +246,7 @@ G_MODULE_EXPORT gboolean common_entry_handler(GtkWidget *widget, gpointer data)
 			dload_val = convert_before_download_f(widget,value);
 			g_signal_handlers_block_by_func (widget,(gpointer) std_entry_handler_f, data);
 			g_signal_handlers_block_by_func (widget,(gpointer) entry_changed_handler_f, data);
-			tmpbuf = g_strdup_printf("%1$.*2$f",dload_val,precision);
+			tmpbuf = g_strdup_printf("%1$.*2$f",(gfloat)dload_val,precision);
 			gtk_entry_set_text(GTK_ENTRY(widget),tmpbuf);
 			g_free(tmpbuf);
 			g_signal_handlers_unblock_by_func (widget,(gpointer) entry_changed_handler_f, data);
